Update README.md
I enjoy watching anime series because they have interesting stories and characters. the first anime i watched was Tokyo Ghoul, and it became my favorite because it introduced me to the world of anime. I love its dark story and the emotions of the main character. The second anime I like is Baki, which my father Recommend to watch it, making it special. The third is Kengan Ashura, Which i enjoy because of its cool and detailed fighting animation.